requirements.txt
pip
Flask
requests
requests_cache
cassandra-driver